requirements.txt
flask==2.0.1
matplotlib==3.5.1
numpy==1.22.0
python-dotenv==0.19.2
requests==2.27.1
python-telegram-bot==13.11
opencv-python==4.5.5.64
pillow==9.0.0 